Change robin_hash.h distance_type to int64_t to prevent overflow loop

+++ pxr/base/tf/pxrTslRobinMap/robin_hash.h
@@ -174,7 +174,7 @@ class bucket_entry : public bucket_entry_hash<StoreHash> {
 
  public:
   using value_type = ValueType;
-  using distance_type = std::int16_t;
+  using distance_type = std::int64_t;
 
   bucket_entry() noexcept
       : bucket_hash(),
